Move URI to data type registry
