+++ ical/types/__init__.py
@@ -6,6 +6,7 @@ from . import boolean, date, date_time, duration  # noqa: F401
 from . import float as float_pkg  # noqa: F401
 from .geo import Geo
 from .period import Period
+from .uri import Uri
 from .utc_offset import UtcOffset
 
-__all__ = ["Period", "Geo", "UtcOffset"]
+__all__ = ["const", "Period", "Geo", "UtcOffset", "Uri"]
